requirements.txt
gradio>=4.0.0
requests>=2.25.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
mcp-sdk[ws]>=0.1.0
anyio>=3.7.0
sse-starlette>=1.6.0
fastapi>=0.100.0
uvicorn>=0.23.0